requirements.txt
flask==1.0
frozen-flask==0.12
beautifulsoup4==4.5.1
PyYAML==5.4.1
